requirements.txt
argparse
netifaces
jinja2
pyyaml
shlex